+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ All notable changes to the kytos project will be documented in this file.
 UNRELEASED - Under development
 ******************************
 
+Added
+=====
+ - Added vlan_pool configuration on kytos.conf to support mef_eline
+
 Changed
 =======